+++ components/Panel/PanelList/EPListFull.tsx
@@ -6,7 +6,7 @@ import { useReducerAtom } from "jotai/react/utils";
 import { panelHistoryAtom, panelReducer } from "@/lib/panel";
 import { EPItemList } from "@/components/EPItem/EPItemList";
 import { Fragment, useEffect } from "react";
-import { LoadMore } from "@/components/LoadMore";
+import { LoadMore } from "@/ui/LoadMore";
 import { useInView } from "react-intersection-observer";
 
 export function EPListFull({ subject_id }: { subject_id: number }) {
